+++ app/AdminModule/presenters/RecordPresenter.php
@@ -8,7 +8,8 @@ class RecordPresenter extends BasePresenter {
 	private $recordManager;
 	/** @var \ITimeLineControlFactory @inject */
     public $timeLineControlFactory;
-
+	/** @persistent */
+	public $backlink = '';
 	protected $data;
 	protected $type;
 
@@ -27,26 +28,49 @@ class RecordPresenter extends BasePresenter {
 
 	public function actionDelete($recordId)
     {
-		$this->recordManager->delete($recordId);
-		$this->flashMessage('stranka byla úspěšně smazána.', 'success');
-        $this->redirect('Record:list');
+		$record = $this->recordManager->get($recordId);
+		if($this->user->isInRole('admin') || $record->uzivatel_id == $this->user->getId())
+		{
+			$this->recordManager->delete($recordId);
+			$this->flashMessage('Záznam byl úspěšně smazán.', 'success');
+			$this->restoreRequest($this->backlink);
+			$this->redirect('Record:list');
+		}
+		$this->flashMessage('Na danou akci nemáte opřávnění.', 'error');
+		$this->restoreRequest($this->backlink);
+		$this->redirect('Record:list');
     }
 
 	public function actionFinish($recordId)
     {
-		$data = array('id' =>$recordId, 'splneno' => 1 );
-		$this->recordManager->edit($data);
-		$this->flashMessage('Úkol byl dokončen.', 'success');
-        $this->redirect('Record:list');
+		$record = $this->recordManager->get($recordId);
+		if($this->user->isInRole('admin') || $record->uzivatel_id == $this->user->getId())
+		{
+			$data = array('id' =>$recordId, 'splneno' => 1 );
+			$this->recordManager->edit($data);
+			$this->flashMessage('Úkol byl dokončen.', 'success');
+			$this->restoreRequest($this->backlink);
+			$this->redirect('Record:list');
+		}
+		$this->flashMessage('Na danou akci nemáte opřávnění.', 'error');
+		$this->restoreRequest($this->backlink);
+		$this->redirect('Record:list');
     }
 
 	public function actionEdit($recordId)
     {
         $record = $this->recordManager->get($recordId);
         if (!$record) {
-            $this->error('Strana nebyla nalezena');
-			$this->redirect('Record:default');
+            $this->error('Záznam nebyl nalezen.');
+			$this->redirect('Record:list');
         }
+		if(!$this->user->isInRole('admin') && $record->uzivatel_id != $this->user->getId())
+		{
+			$this->flashMessage('Na danou akci nemáte opřávnění.', 'error');
+			$this->restoreRequest($this->backlink);
+			$this->redirect('Record:list');
+		}
+
 		$record = $record->toArray();
 		$date = new \DateTime($record['datum']);
 		$record['datum'] = $date->format('d F, Y');
@@ -65,6 +89,17 @@ class RecordPresenter extends BasePresenter {
 		{
 			$this->template->records = $this->recordManager->getUserRecord($this->user->getId())->order(\App\Model\RecordManager::COLUMN_DATE.' DESC');
 		}
+	}
+
+	/**
+	 * Render record detail.
+	 * @param type $id
+	 */
+	public function renderDetail($id)
+	{
+		$record = $this->recordManager->get($id);
+		$this->template->record = $record;
+		$this->template->userRef = $record->ref('uzivatel');
 	}
 
 	public function renderList($own = null)
@@ -86,10 +121,10 @@ class RecordPresenter extends BasePresenter {
 		$form->addHidden('id', null);
 		$form->addHidden('uzivatel_id', null);
         $form->addText('nazev', 'Název:')
-            ->setRequired();
+            ->setRequired('Zadejte název.');
 
 		$form->addTextArea('popis', 'Popis:')
-            ->setRequired()
+            ->setRequired('Zadejte popis.')
 			->getControlPrototype()->setId('editor');
 
 		$form->addText('datum', 'Datum:', 30, 30)->getControlPrototype()->setClass('datepicker');
@@ -109,13 +144,20 @@ class RecordPresenter extends BasePresenter {
 		$date = new \DateTime($values['datum']);
 		$values['datum'] = $date->format('Y-m-d H:i:s');
 		$date = new \DateTime($values['datum_splneni']);
-		$values['datum_splneni'] = $date->format('Y-m-d H:i:s');
+		if($values['datum_splneni'])
+		{
+			$values['datum_splneni'] = $date->format('Y-m-d H:i:s');
+		}
+		else
+		{
+			unset ($values['datum_splneni']);
+		}
 		if ($values['id']) {
 			$this->recordManager->edit($values);
         } else {
             $this->recordManager->add($values);
         }
-        $this->flashMessage('Partner byl vlozen', 'success');
+        $this->flashMessage('Záznam byl vložen', 'success');
         $this->redirect('Record:list');
     }
 }
